🚀 Deploy beta

+++ test/utils/ConsolidatedEvents.sol
@@ -54,7 +54,7 @@ contract ConsolidatedEvents {
         uint256 conditionalOrderId,
         uint256 fillPrice,
         uint256 keeperFee,
-        uint kwentaFee
+        uint256 kwentaFee
     );
     event FeeImposed(
         address account, uint256 amount, bytes32 marketKey, bytes32 reason
